requirements.txt
discord.py
gtts